pyproject.toml
[project]
name = "fimfarchive"
version = "0.1.0"
description = "Preserves stories from Fimfiction"
readme = "README.md"
requires-python = "~=3.8"
dependencies = [
    "arrow~=1.3",
    "bbcode~=1.1",
    "blinker~=1.8",
    "jinja2~=3.1",
    "jmespath~=1.0",
    "jsonapi-client",
    "pillow~=10.4.0",
    "requests~=2.32",
    "tqdm~=4.66",
]

[project.optional-dependencies]
lz4 = ["lz4~=4.3"]

[tool.uv]
compile-bytecode = true
dev-dependencies = [
    "mypy~=1.11",
    "pytest~=8.3",
    "requests-mock~=1.12",
    "ruff~=0.6",
    "types-jmespath~=1.0",
    "types-requests~=2.32",
    "types-tqdm~=4.66",
]

[tool.uv.sources.jsonapi-client]
git = "https://github.com/JockeTF/jsonapi-client.git"

[tool.mypy]
files = ["fimfarchive", "tests"]
ignore_missing_imports = true

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
extend-exclude = ["fimfarchive/__init__.py"]

[tool.ruff.lint]
select = ["A", "E", "F", "N", "RUF", "T10", "W"]
ignore = ["N801", "RUF012", "RUF022"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
